extract raw-files.js generator out of Gruntfile into separate module
